[NEAR-119] Refactor: Concert 모델 분리
